Added index file and timeout for initial fetch from local storage
